Move kiwix icon to app module
