+++ app/src/main/java/org/kiwix/kiwixmobile/webserver/wifi_hotspot/HotspotNotificationManager.java
@@ -28,7 +28,7 @@ import android.os.Build;
 import androidx.annotation.NonNull;
 import androidx.core.app.NotificationCompat;
 import javax.inject.Inject;
-import org.kiwix.kiwixmobile.core.R;
+import org.kiwix.kiwixmobile.R;
 import org.kiwix.kiwixmobile.core.utils.Constants;
 import org.kiwix.kiwixmobile.webserver.ZimHostActivity;
